// client/src/app/models/Students.ts
import { Title } from '@angular/platform-browser';

export interface Students{

    TipoIdentificacion?: string;
    Identificacion?: string;
    Nombre1?: string;
    Nombre2?: string;
    Apellido1?: string;
    Apellido2?: string;
    Email?: string;
    Celular?: string;
    Direccion?: string;
    Ciudad?: string;

}